requirements.txt
fastapi
pillow
numpy
gradio
toml
curl_cffi
snowflake-id
discord